+++ src/options.ml
@@ -34,12 +34,13 @@ let execution_mode = ref WithSideEffect
 let sorting_policy = ref OnTheFly
 
 
-let opt_lsp   = ref false
-let opt_json  = ref false
-let opt_pt    = ref false
-let opt_ext   = ref false
-let opt_ast   = ref false
-let opt_typed = ref false
+let opt_lsp     = ref false
+let opt_service = ref false
+let opt_json    = ref false
+let opt_pt      = ref false
+let opt_ext     = ref false
+let opt_ast     = ref false
+let opt_typed   = ref false
 let opt_all_parenthesis = ref false
 let opt_nse   = ref false
 let opt_sa    = ref false