+++ src/tools.ml
@@ -82,6 +82,7 @@ module Option : sig
   val get_exn     : exn -> 'a option -> 'a
   val get_dfl     : 'a -> 'a option -> 'a
   val get_fdfl    : (unit -> 'a) -> 'a option -> 'a
+  val get_list    : ('a option) list -> 'a list
   val iter        : ('a -> unit) -> 'a option -> unit
   val map         : ('a -> 'b) -> 'a option -> 'b option
   val bind        : ('a -> 'b option) -> 'a option -> 'b option
@@ -112,6 +113,12 @@ end = struct
 
   let get_fdfl dfl =
     function None -> dfl () | Some e -> e
+
+  let get_list l =
+    List.fold_right (fun x accu ->
+        match x with
+        | Some v -> v::accu
+        | None -> accu) l []
 
   let get_all xs =
     let module E = struct exception Aaarg end in